Start removing code supporting async action execution.
This code removes support for the case when actions return a continuation. This
removes one layer of the state machine in SkyframeActionExecutor; the other
one, ActionStepOrResult, must stay because action postprocessing can result in
a Skyframe restart if the action discovers new inputs and it would be truly
awful if we had to re-run the whole action in that case.

RELNOTES: None.
PiperOrigin-RevId: 490216488
Change-Id: Icf8567ca54c5301c47f4effa7b6399582882653a