Call notifyActionCompletion() in fewer places.
This is now possible because executeAction() does not need to be prepared for
the case that the action is not actually executed when it returns.

The change in the test case is needed because this changes the order of
ActionCompletionEvent and ActionExecutedEvent for actions that fail. I verified
that this does not result in a visible change of the output regardless of the
value of --keep_going .

Also refactor the mechanics of signaling lost inputs on the side.

RELNOTES: None.
PiperOrigin-RevId: 490216945
Change-Id: Id587d8f6bde391eeb94ece4ad382ceb7805f373c