Subtract imported_library in subtractSubtrees
This is needed for some upcoming changes to migrate linking info to CcInfo.
During the migration we will put the linking info in both CcInfo and
ObjcProvider, and it is possible in some code path (through swift_library, fwiw)
for an imported_library to become a regular library.  For avoid_deps to work,
avoid_deps needs to add imported_library to the list of libraries to subtract.

PiperOrigin-RevId: 490275745
Change-Id: I013950a2ffe988013ca1613d8b7802ca379d3170