Add auto-configured Unix toolchain support for `BAZEL_CONLYOPTS`
Since `BAZEL_CXXOPTS` is only used for C++ compilation, there was no
way to pass in options for C compilation that would also be used when
discovering built-in include directories with `-xc`.

When using the non-Xcode auto-configured toolchain on macOS, it can be
necessary to specify the macOS SDK sysroot via a compiler flag. While
this can usually be achieved by combining `--conlyopt` with
`BAZEL_CXXOPTS`, this assumes that the built-in include directories
returned with `-xc` are always a subset of those returned with `-xc++`.
Having `BAZEL_CONLYOPTS` would allow this situation to be navigated in
a way that is clearly correct.