Introduce NoConfigTransition for rules that don't need configs.
 - Provides an alternative for rules that use the host transition as a proxy for
   nonconfigurable targets. Host transition does that only as a side effect,
   and isn't the right tool for that purpose
 - Unblocks host config logic removal
 - Replace environment()'s host transition

NoConfigTransition returns a mostly empty BuildOptions: it has CoreOptions but nothing else. Ideally it'd be completely empty. But too much core Bazel logic reads CoreOptions. We should remove those dependencies. But that's a larger task. See comments in PR diffs for details.

This is very similar to NullTransition. This PR doesn't try to merge them because of deep Bazel invariants that NullTransition == source file. Perhaps that can be relaxed. But that's also a larger task.

PiperOrigin-RevId: 490598476
Change-Id: Ic4dedc37659d7dc932d7bedc99db17a948566169