Remove calls to SpawnStrategy.beginExecution() from StandaloneTestStrategy.
This requires some code duplication and I'm fairly sure the code can be
formulated in a better way, but let's wait with that until the removal of
TestAttemptContinuation is on the menu; it's complicated on both sides
(implementation and call sites) so I'd rather to mechanical transmutations
of the code when that code is not the specific focus.

RELNOTES: None.
PiperOrigin-RevId: 490842244
Change-Id: If5c66d2b0987235d86ef4985e529e641abe34a3e