Refactor test running away from continuation passing style.
This required a heavy refactoring of test cancellation because it used to be
tightly coupled to the exact way the spawn continuation is executed.

It's very possible that not everything is perfect now, but the control flow is
now much simpler so it should be easier to weed out any remaining bugs.

RELNOTES: None.
PiperOrigin-RevId: 491276670
Change-Id: Icb96e0f8fd65af9eb9a7034540c1e8079a6e9e03