Remove a restart from BazelTestAttemptContinuation.
It was a remnant of the time when Spawns were executed in a continuation-
passing style and when I last touched this code, I didn't quite understand it
so I did only a mechanical refactoring. This resulted in code that looked very
illogical (but was provably idempotent to its previous state)

Now, after a closer look, it turns out that the code did not only look silly,
it also *was* silly.

RELNOTES: None.
PiperOrigin-RevId: 491639280
Change-Id: I06b2a231fd95096ad7697c08bc6017275f532625