python: Make python/providers.bzl values the canonical symbols
The builtins injection overrriding that exports.bzl supports doesn't work
very well for providers because the code within builtins isn't able to
refer to the effective value of an injected key. Stated another way, for
providers, the "+/-" setting in exports has to match what the object that
builtins code `load()`s.

Instead of trying to coordinate the two, just make the injected key always
active and point to the symbol in the bzl file, which determines which
implementation of a provider is used. Builtins code and exports both load
that name, so they're always in sync.

Work towards #15897

PiperOrigin-RevId: 491733075
Change-Id: I2b52522320d77996cfdd679baac1c8471ab116fb