Delete experimental_persistent_javac flag
We're looking to cut back on the number of flags, this one doesn't see
much/any use besides https://github.com/google/j2cl/blob/master/.bazelrc,
which can easily be replaced with a config expansion.

At the same time we're looking to improve the introspectability of
commandlines, specifically around workers, which this bazel-internal expansion
obscures.

PiperOrigin-RevId: 492481856
Change-Id: I27f28a3fd4d0d90c86d9c661ed7c9312cf851644