Enable revised output directory hash suffix computation
Flips `--experimental_output_directory_naming_scheme` and
`--experimental_exec_configuration_distinguisher` to
`diff_against_baseline` and `off`, respectively, which makes it so that
Starlark transitions can return to previous configurations, thus
improving cache hit rates.

Also changes the fixed platform suffix in the exec configuration from an
empty string to `"exec"` for compatibility with logic that looks for the
 substring `-exec-` in paths to determine whether the current
configuration is the exec configuration.

As a result of this commit, output directory paths can change in the
following ways:
* `-ST-<hash>` suffixes can change or disappear and, in rare cases,
  appear where they previously didn't
* for exec configuration directories, `k8-opt-exec-2B5CBBC6` becomes
  `k8-opt-exec-ST-011b9bdc32ed`

This may result in tool paths that are seven characters longer, which
can cause builds on Windows to fail that were already very close to the
`MAX_PATH` limit. A follow-up commit will reduce the length of the hash
by three characters, bringing that increase down to four.

Work towards #14023