Don't let AbstractSpawnStrategy cache input mappings of remote actions
In the previous change we made sure that only a bounded number of Merkle
trees remain in memory. This has already cut down worst-case memory
usage somewhat, but it looks like we have the same issue with the the
protocol independent input mapping data structure. That one is used as
an input for the Merkle tree computation, and is memoized by
AbstractSpawnStrategy.

This change adds a willAccessRepeatedly flag to getInputMapping(). This
flag can be used by the caller to indicate whether there is a realistic
expectency that input mappings are going to be used in a repeated
fashion. For all of the cases leading up to REv2's GetActionResult(), we
ensure that we set willAccessRepeatedly to false.

Issue: BEN-14271