Remove CTaD from ConfiguredAspectFactory and CTFactory.createAspect.
The motivation is to cut CTFunction's dependency on Targets of its
dependencies. The plan to achieve this uses ConfiguredTargetAndData
(CTaD) as an abstraction layer. CTaD will no longer directly expose the
underlying target and instead hide information behind API calls.

ConfiguredAspectFactory previously depended on CTaD for direct access to
the Target. There are Target fields used for aspect creation that are
not needed more broadly. So instead of depending on CTaD, this change
makes ConfiguredAspectFactory depend on the broken down ConfiguredTarget
and Target components.

PiperOrigin-RevId: 493189240
Change-Id: Ie255321a2383d9ee06cb64121fce7799d4b65cce