Change the way linkstamp headers's digest is computed
The previous mechanism relied on fingerprinting of nested sets which isn't available in Starlark. Since there are few targets with linkstamps and these targets most of the time do not have many headers, it should not cause a performance regression to flatten the nested set and compute the nested set that way.

No regression was reported in a performance benchmark.

RELNOTES:none
PiperOrigin-RevId: 494161019
Change-Id: If60eb685054a3e03b0450402093e023460bb4596