[site] Remove ios-app tutorial and redirect to rules_apple repo
There have been talks about moving the ios-app tutorial outside the Bazel repo so that it's easier to iterate on and keep it up to date. A few weeks ago the tutorial now living in [rules_apple](https://github.com/bazelbuild/rules_apple) was [revamped](https://github.com/bazelbuild/rules_apple/pull/1639). The corresponding example was also updated in https://github.com/bazelbuild/examples/pull/258.

In the Bazel site there are a few links that point to the ios-app page. This is why I opted for keeping the page in the site not to break links, but I'd be happier if we could simply provide a redirect in the Bazel site to the rules_apple tutorial text now living here: https://github.com/bazelbuild/rules_apple/blob/master/doc/tutorials/ios-app.md

Some previous conversations:
- https://github.com/bazelbuild/bazel/pull/14499
- https://github.com/bazelbuild/examples/pull/237

cc: @aiuto

Closes #16985.

PiperOrigin-RevId: 494660058
Change-Id: I7480be9798585e8523cb0cd3fe2993062aa38e7c