Rewrite def_parser computed default
In Starlark we have been breaking these cycles whenever they happen with tags placed on the involved targets. We can break the cycle in this CL for files in tools/cpp but files in third_party/def_parser aren't in the Bazel repository. For now we break the cycle by checking the name of the target hoping that it is unique enough. But separately the tag will be added to the targets in third_party/def_parser.

RELNOTES:none
PiperOrigin-RevId: 494663707
Change-Id: Iad498ef4f5981e3cf766d86d99b744b902d47711