Fix runtime classpath computation in java_binary/java_test
The correct order is in `JavaInfo.compilation_info.runtime_classpath` and not `JavaInfo.transitive_runtime_jars`. But the former is lost after we merge the generated extension registry `JavaInfo`, so this change reorders the code a little to make sure we can compute the correctly ordered classpath.

Additionally, we make the runtime/compile-time deps collection identical to native

PiperOrigin-RevId: 494673627
Change-Id: Iea45cefd30a1e559e8098203da35afd98762d457