Make launcher resolution in java_binary/java_test identical to native
In native, the `$java_launcher` attribute gets set to the `launcher` attribute if the latter is set on the rule.
Before this change, in Starlark, the `_java_launcher` attribute is a configuration_field if launcher flag must be resolved, and None otherwise.
This change makes it a computed default on the `launcher` attribute. Unfortunately, this means we need another rule class since input attributes to a computed default can not be None.

PiperOrigin-RevId: 494674154
Change-Id: I8e9e4e6a4f45050cc910012e43541d4dc78028be