Add --low_priority_bazelrc
This startup flag behaves exactly as --bazelrc does, but is applied
before the workspace bazelrc rather than after.

This is useful for people who write and distribute bazel wrappers (which
appears to be most large companies). This allows a bazel wrapper to
supply new _default_ values to flags, which a repo may choose to
override.

Without this change, it's impossible for a bazel wrapper to decide to do
something like set `--incompatible_enable_cc_toolchain_resolution`
because they have no practical way of doing so. if they set this on the
command line or in a .bazelrc file passed to the `--bazelrc` flag, it
cannot be overridden other than on the command line. This flag allows a
wrapper to point at a bazelrc file with extra defaults, which can be
overridden in the repo bazelrc file.