Include full tree artifact in inputs when prefetcher doesn't support partial tree artifacts.
The actions generated by SpawnActionTemplate and CppCompileActionTemplate
contain individual TreeFileArtifacts, but not the respective TreeArtifact, in
their inputs. This causes the input prefetcher to crash when building without
the bytes, as it is only able to fetch entire tree artifacts and expects their
input metadata to be available.

This PR introduces an ActionInputPrefetcher#supportsPartialTreeArtifactInputs
method signaling whether prefetching partial tree artifacts is supported; if
not, the entire tree artifact is included in the inputs of any such actions.

This is a temporary workaround to unblock the 6.0.0 release. The long-term plan
is to fix #16333 to enable prefetching partial tree artifacts, and remove this
workaround.

Fixes #16987.