Pack the last changed and last evaluated versions in `InMemoryNodeEntry` into a single field of type `NodeVersion`.
This opens up a free field to add `SkyKey` to `InMemoryNodeEntry` (a 6-field object has the same cost as a 7-field object).

In the common case that a node was last evaluated and changed at the same version, use that version to implement `NodeVersion`. In the less common case of change pruning, use a wrapper.

PiperOrigin-RevId: 495691444
Change-Id: Ibeb09c1d0546d0ac053690aeccd69f05e88eee8c