Fix bug in benchmark stats
The number of requested iterations was accumulating every time the timer was stopped, instead of every time a benchmark run completed. This meant that tests which called b.stop() explicitly, particularly inside a loop, would have drastically inflated figures for the number of iterations and report artificially high performance per iteration.

Discovered while benchmarking for #13179.

Also made more room for the benchmark name in the displayed table, and added a benchmark for list appending without reallocation. To compare the latter to list appending with reallocation, you have to invert the test run order (by renaming the test cases) to account for warmup effect. (I found them to be basically indistinguishable, indicating that reallocations are neglible compared to the Starlark loop evaluation).

PiperOrigin-RevId: 495885589
Change-Id: I57b54ca0b4439109c5659a5ae08cdf10bec3114f