Bazel DevSite: Fix links that rely on redirects.
During the recent site overhaul a lot of files were moved to new locations, without updating any links to them. Instead, redirects were set up.

This works fine for files @ HEAD, but breaks horribly for the 6.0 docs (and any future release docs).

Consequently, this CL fixes the links for both HEAD and 6.0 to avoid such problems in the future.

PiperOrigin-RevId: 496481366
Change-Id: I0bfde556cbb61d6ba50dd24120febf87ab46302d