[Skymeld] Collect BuildGraphMetrics.
This CL adds a feature that was missing in Skymeld: the collection of BuildGraphMetrics.

This is done via a traversal of the ActionLookupValues when all analysis work is finished. Though CPU-bound, it's not expected to be on the critical path of the build.

There's a difference in behavior here between Skymeld and non-Skymeld:
- Non-Skymeld: this metric is not collected when no conflict checking (and no ALVs traversal) is done.
- Skymeld: this is done regardless, as part of the analysis finished callback.

PiperOrigin-RevId: 496896389
Change-Id: I2d06ffaadb0fa2786a179569f32ec97780bd0626