Prevent change-pruning for `ActionExecutionValue` when the `discoveredModules` have changed.
It is possible for a change to just the C++ features to add/remove actions for the corresponding configured target, which in turn affects the generating actions for the artifact (different action index). When than happens, the artifacts in an old `ActionExecutionValue::discoveredModules` may become stale. Today, in absence of any other change, such an `ActionExecutionValue` is subject to change-pruning leaving the value with stale `ActionLookupData`. That later will crash when a downstream action tries to use a discovered module with such artifact.

Use `NestedSet::shallowEquals` which is potentially overly sensitive to avoid materializing the nested set just for the sake of comparison. This seems like a reasonable tradeoff since:
  1. The output `.pcm` file is already quite sensitive to changes in transitive headers, making header edits in a dep affect the `ActionExecutionValue` for the module already.
  2. Pure changes to C++ features are typically done via a change which affects the package, which affect the configured targets transitively (`RuleConfiguredTargetValue` has transitive packages and identity equality), this marks all actions being as `NEEDS_REBUILDING` anyway.

PiperOrigin-RevId: 499254933
Change-Id: I88abaf62f767fee4c03fe1e108312cbff2b5f494