WIP: Mac Catalyst support for bazel / rules_swift / rules_apple
This is the first of a few PRs to add Mac Catalyst support for Bazel. I
have an app that can be successfully built with a few manual patches to
the bundle (proper actool usage, different bundle Info.plist, codesign,
and notarization). But overall, it can be successful built with an app
that passes macOS AppStore review / automatic checks.

This is WIP, as I need to migrate a few manual input flags into
crosstool section, and need advices from Bazel maintainers.