Rollback #14510 because it causes remote test execution to fail
When the parent process is run as a different user to the test
itself (as can happen in some remote build environments, such
as buildbarn) using `kill -0 $PPID` fails, causing the test to
be prematurely killed.

This PR rolls back the (nicer, cleaner, not always working)
fix and replaces it with the (less nice, less clean, more often
working) check that was there before.

Tested with a local build against buildbarn.