Remove references to HostTransition.INSTANCE
This updates a lot of tests that rely on the host transition. Tests are updated in one of three ways:

- remove the test case (it's explicitly testing "host", which is no longer relevant)
- replace host references with exec
- "host" is a stand-in for "any transition", so replace with an arbitrary transition

There's some duplication in the test infrastructure logic for providing exec configurations. This is because different tests inherit different base infrastructure and not all symbols (i.e. AnalysisTestUtil) are equally accessible everywhere.

PiperOrigin-RevId: 499531378
Change-Id: Ia16e73a7ef56b642bd8143072fdc9df395516bdd