Rollback #14510 because it causes remote test execution to fail
When the parent process is run as a different user to the test itself (as can happen in some remote build environments, such as buildbarn) using `kill -0 $PPID` fails, causing the test to be prematurely killed.

This PR rolls back the (nicer, cleaner, not always working) fix and replaces it with the (less nice, less clean, more often working) check that was there before.

Tested with a local build against buildbarn.

Closes #17147.

PiperOrigin-RevId: 500691976
Change-Id: Ife55db5c1ed88d40ba502257ce9b679b04eeb179