Remove --apple_platforms applying to host
If you're targeting an Apple platform like
`--apple_platforms=:ios_something`, you don't want this to apply to the
host configuration, since that implies your host would be iOS as well.
Today this was ignored because `applePlatformType` is hard coded to
macOS a few lines up, but this still affects the configuration in a way
you wouldn't expect.