Remove RANGE requests on retry
Addresses review comment
https://github.com/bazelbuild/bazel/pull/14258#discussion_r1049719234 .

RANGE requests are an optimization in that they avoid resending
previously sent data over the wire. However, they are not necessary to
implement retries with the HTTP remote cache.

The implementation so far was only correct for uncompressed streams, for
compressed streams the offset has to be set in terms of encoded bytes
instead of decoded bytes.