Implement GroupedList variant to track deps in a HashSet
This variant is used in InMemoryNodeEntry construction when a node's
corresponding key opts into Partial Reevaluation.

Doing so avoids O(|deps|) work each time such a node is reevaluated.
Partial Reevaluation nodes may be reevaluated many more times than
"regular" nodes.

The O(|deps|) work this avoids is constructing the
SkyFunctionEnvironment.PartialReevaluation.previouslyRequestedDepsSet on
every reevaluation. With the new variant GroupedList, this set is
maintained in the GroupedList as it mutates, and the environment
references the set.

Having this set is essential because the Environment needs to keep track
of whether requested deps are new (as of the given node reevaluation, not
as of the Skyframe evaluation version, to be clear),

The memory cost of deps is very important, so this functionality avoids
growing the memory layout of regular GroupedList objects. We expect there
to be many many fewer "Partial Reevaluation" nodes than regular nodes so
having a larger memory cost for them is less important.

(At the time of this CL, the only Partial Reevaluation nodes correspond
to the scopes of genquery rules. The number of unique scopes in a build
is likely fewer than 50.)

Preserving a GroupedList's preference for HashSet-based dep tracking
through "dep compression" is not necessary, given how compression is
currently used. Done nodes' deps get compressed, losing the GroupedList
wrapper. When those done nodes return to an "inflight" status, their
compressed form gets "rehydrated" into a GroupedList again, and stored in
a DirtyBuildingState. In particular, "Partial Reevaluation" nodes do not
require efficient containment checks for GroupedLists in
DirtyBuildingStates -- indeed, nothing performs containment checks on
those deps.

PiperOrigin-RevId: 501044175
Change-Id: I8ff79089b8aac362adcb2661371f1bf620f37ac4