Start fresh ActionResult download on retry
Addresses review comment
https://github.com/bazelbuild/bazel/pull/14258#issuecomment-1352743891

Continuing to stream from the previously read bytes on retry is safe for
CAS downloads because items stored in the CAS can't change (otherwise it
would change their content address). However, the same is not true for
action cache items. Therefore, we need to start a fresh download on
retry for action cache items.

To that end we need to lift the retry logic out of the `get` utility
function that implements the details of the download, to the separate
higher level blob and action result download functions.

The counter of bytes downloaded also needs to be lifted and is turned
into an optional parameter to double as the flag indicating whether the
download is a CAS download or not.