Build with incompatible_disallow_empty_glob
In order to flip the flag, all downstream projects should be adapted. However, it is hard to fix them all if there are constant regressions. Adding it to the CI will ensure that once the project can build with incompatible_disallow_empty_glob it can keep building like that.
See: bazelbuild/bazel#15327

Closes #16670.

PiperOrigin-RevId: 501764139
Change-Id: I909a08364794bc657aa6fefce4cf094a10563571