Self transition constraint_setting and constraint_value to the NoConfig state.
These rules can only define constants, and all their attributes are not configurable.

This reduces the number of configurations constraint values appear in, but
does yet reduce that to a single instance per value. For this cquery

```
blaze cquery --compilation_mode=opt --fat_apk_cpu=x86 \
  --android_platforms=//buildenv/platforms/android:x86 \
  --incompatible_enable_android_toolchain_resolution=1 \
  --//third_party/java/android:system_api=true \
  'allpaths(//java/com/android/dialer:dialer, //third_party/bazel_platforms/...)' \
  | grep bazel_platforms/os | sort -u
```

We see 71 different combinations of constraint values and cquery configs
before this change and 37 after.  That is better, but inexplicably much
higher than the 8 or so I would expect.

RELNOTES: None
PiperOrigin-RevId: 501909093
Change-Id: Ie66ac2b144837050fdb3e1fd18f5dc1ba155e480