Generate docs for rules referenced in structs.
We have rules that are directly referenced via a struct and do not have a starlark function wrapping the rule. These currently do not generate documentation, while other struct referenced fields do if they refer to a starlark function. This CL adds support for generating docs for rules which are referenced in structs.

PiperOrigin-RevId: 502977318
Change-Id: Iaf7fab8c3fc5c2685727193abbc42e5e3a13c57b