Remove nocopts attribute from cc_library.bzl
Using this attribute was disabled, see #8706. It was added to the Starlark rule definition although it wasn't working. It is now moved to semantics while we do the necessary internal migration to remove it completely.

output_filter_test doesn't make sense in Bazel since nocopts is not available. The file must remain while we migrate internally away from nocopts though.

RELNOTES:none
PiperOrigin-RevId: 503155992
Change-Id: I43184941c9cb9ee97d5229cf37062b346643a48a