Allow tree artifacts in source attributes to have arbitrary extensions
Currently, tree artifacts in srcs/hdrs/non_arc_srcs need to have extensions that
match their contents, which is unnecessary and silly.  The new behavior mirrors
how cc_library behaves.

This change also imports a check in cc_library that checks that a source file is
not duplicated -- this may cause blaze to error out when it previously wouldn't.
The fix is simply to deduplicate the list of source files.

This requires way more code factoring that it should, but the final result gets
objc_library closer to cc_library and hopefully is easier to maintain.  We were
able to add functionality while deleting more code than we add (if we don't
count tests).

PiperOrigin-RevId: 503189010
Change-Id: I9a20a71753183527b7c0391b94968d7acd6de1a4