python: Skip tests not supported by the Google Starlark rule implementation.
This facilitates deletion of Google-specific Java code for the Python rules.

The skipped tests all relate to Python 2 support, which the Google rules have
no support for. Since this also isn't implemented in the Bazel Starlark rules,
the overrides are left in place for that case.

Also enables the test for toolchain resolution with the Starlark rule
implementation since that feature has been implemented.

PiperOrigin-RevId: 503473007
Change-Id: I57feb0e1d12bc8b4d48d686684eeb8db959fd66d