Small tweaks to platform documentation. - Point platform rules to the explanatory document which people should have read first. - Update the left nav to include the word "Toolchains" - move https://bazel.build/extending/platforms from Design Docs to Concepts
It's not much, but it is a start.

Note to the team: The automatic docs from source system makes this painfully complex. We need a way to generate rule doc from source but keep all the organization of how we group rules under what headings to the framework on the docs serving side.

Next steps:
- Update the platforms page.

RELNOTES: None
PiperOrigin-RevId: 504460917
Change-Id: I0acc1bd9ef8f6381f5c898688dddf328dcf513aa