Fix a bad change to the bazel encyclopedia rules template.
I thought I was editing one instance, not the template for everyone.
I did not expect that I would have to put text like this in the Java code.
That is too difficult for anyone except the team. https://github.com/bazelbuild/bazel/issues/17332

Also, fix https://bazel.build/reference/be/overview so that the TOC
in the center of the page agrees with the words on the left nav.
Now, *why* this is the 6.0.0. _toc.yaml is a mystery.  That seems
like a bad indirection through _buttons.html, where the version
is getting added. No point in figuring it out right now. The change
is correct for 6.x as well.

RELNOTES: None
PiperOrigin-RevId: 504836140
Change-Id: I572170fdd7dd48d43710e28dc3d2e75ae4f5d065