Allow Java coverage collection for external targets.
A roll-forward of https://github.com/bazelbuild/bazel/pull/16268.

LazyWritePathsFileAction has been changed to accept a custom converter
parameter that can be used to specify exactly what path should be
written out to the file. This is required to support the use case of an
internal rule that still needs root-relative paths written out.

The default case (when no converter is specified) is to output the exec
path, as per the original PR.

PiperOrigin-RevId: 505678128
Change-Id: Ib1d8547dcb359a9e0fb7e462850424ec4218f7fb