documentation: clarify combined cache
In Bazel, a Combined Cache is in effect when both Local Disk Cache
(--disk_cache=) and Remote Cache (--remote_cache=) are set.
This concept, however, only exists in Bazel source code and not defined
anywhere in Bazel's documentation.

Let's clarify Combined Cache concept in the doc to explain to users
what it is.