Test and fix root symlink edge case in runfiles library
Due to incomplete test coverage for the case of a root symlink named like a Bazel module, the Bash runfiles libraries behaved incorrectly in this case. This commit fixes that and adds proper test coverage for all libraries by adding repo mappings that conflict with the existing rlocation testcase for `config.json`.

Closes #17317.

PiperOrigin-RevId: 505873412
Change-Id: I72019d329b72cab9cf893deb714da8889d371617