Use WeakInterner to hold `LibraryToLinkValue`s in CcToolchainVariables
Previously, different `LibraryToLinkValue` inherited classes are created so that some heap overhead can be reduced. However, it is still possible equal objects with the same `name` are constructed multiple times, leading to multiple copies of the same object in heap.

The most straightforward idea to reduce the overhead is to utilize `Interner`. So, this commit applies `WeakInterner` to all of these static creation methods.

PiperOrigin-RevId: 505987062
Change-Id: I9e99adb3b2a6788467f6a290fb4fcaecf8ff0b65