Add shared actions to the execution graph.
Currently, only the action that wins the execution race is represented in the execution graph. Dependency edges on actions which lose the shared action execution race are not represented. This could result in an underestimated critical path.

With this change, shared actions all share the same execution graph node. Artifacts from the executed action are guaranteed to already exist in the execution graph because `ActionCompletionEvent` is posted prior to construction of `ActionStepOrResult.Finished`, which is needed for the shared action to observe it as done. Nevertheless, I used `logUnexpected` to avoid a crash if this assumption does not hold.

PiperOrigin-RevId: 506140972
Change-Id: Ia2d5e48c524a1afab198d3c61f21a5d893aa1cee