Make WithDriver the default for ValueOrExceptionProducer.
Removes the option to use this without a Driver instance.

In practice, it's better to have parent state machines forget about
their children as early as possible to reduce memory pressure. The
ValueOrException API forces parent machines to keep a reference to
their children for querying results after they complete. Instead, the
parent can inject a consumer-like callback into the child for receiving
results and not keep child references. Children are implicitly forgotten
on completion.

At the SkyFunction level, implicitly dropping references on completion
is unimportant because the SkyKeyComputeState must keep a reference to
the Driver to resume the computation. It would therefore would need to
explicitly drop the Driver reference anyway.

PiperOrigin-RevId: 506422024
Change-Id: I8d5444045f0ab1ea47a7170666b0a78927b0bd9a