Refactor ConfiguredTargetFunction into two classes.
- A new class - PrerequisiteProducer - handles analysis through dependency evaluation.
- ConfiguredTargetFunction then runs rule implementation logic and returns final results.

This should be a no-op for both semantics and performance. It should apply the same exact logical sequence as the original code. This is important because ConfiguredTargetFunction implements a crucial, performance-sensitive phase of the build process that's very sensitive to subtle logical regressions. So this change is *not* trying to optimize anything. It's purely a refactoring of where code lives.

The motivation for this change is to support injecting a new code path that trims configurations to reduce graph bloat. But it's also intended to stand on its own in case trimming experiments ultimately don't show value. The other arguments are:

- ConfiguredTargetFunction is a massive, complex file. Breaking it down helps limit conceptual scope.
- This enforces stronger APIs. This is useful because other parts of the code base - particularly AspectFunction - call directly into various ConfiguredTarget methods, with slightly different parameters, basically bypassing all abstractions. The logical intermixing produces really delicate code with non-trivial maintenance burden. I believe these new abstractions mitigate that at least a little bit.
- This could support a more resilient cquery that doesn't fail when rule implementation logic fails. This is an annoying problem today: users try to figure out why their build fails by cquerying over it, but cquery itself fails so it's a dead end as a diagnostic tool. This is unnecessary since cquery doesn't need to run rule implementation logic.

I acknowledge the breakdown isn't even. I was hoping to be able to extract out ~50% of ConfiguredTargetFunction (and chop any one class's conceptual complexity in half). In reality, this turns a 1369-line ConfiguredTargetFunction into a 377-line ConfiguredTargetFunction and 1201-line PrerequisiteProducer. Some of that is new comments and boilerplate Java. But still: not an even break.

The core change here forks ConfiguredTargetFunction into PrerequisiteProducer. Other files are small reference changes or formerly inner classes that graduated to full files. Almost all code is simply moved, not new code.

PiperOrigin-RevId: 506489649
Change-Id: I42cfe050cd7a43948855b61ecc7fe3ff1f52d68a